Add files via upload
Explination:
Step 1:Import Libearies-I am using Scanner class from the java.util package.Scanner class is used to obtain user input.
Step 2:Class Declaration-Than I declare a class named SimpleBankingApplication.This class contains the main method and other methods to perform banking operations.
Step 3:Main Method-It starts the execution of the application.This is the entry point of the program.
step 4:Menu Display and User Choice-Inside the main method,I am using while loop that runs as long as isRunning.It displays a menu of banking operations and waits for the user to input a choice. and also used switch statment fordeposit(), withdraw(), checkBalance(), or exits the application.
step 5:Menu Printing Method-This method prints out the menu options for the user to choose from.
step 6:Deposit Money in bank command- the user to enter an amount to deposit money in bank.If the amount is valid (greater than 0), it adds it to the balance. Otherwise, it prompts the user to try again.
step 7:-Withdraw money in bank-the user to enter an amount to withdrawin bank.It checks if the amount is valid (greater than 0) and if there are sufficient funds in the account. If both conditions are met, it deducts the amount from the balance.
step 8:-Check Balance in bank- simply prints out the current account balance.
step 9:-exit command-then close the application.

It implements a simple command-line banking application allowing users to deposit, withdraw, and check their balance.